Use preserve_most calling conventions for Zone::NewExpand
The zone uses bump allocation in the fast-path, and falls back to
allocating a new segment if there is not enough space.
Since this is rarely executed and zone allocations happen a lot, we
should mark `Zone::NewExpand` as "preserve_most" to make
`Zone::Allocate` as fast and slim as possible.

R=bikineev@chromium.org, leszeks@chromium.org

Change-Id: I0d592a35440bc3d61ca04425fc2f98c8a8bbbaae
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4146436
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#85258}